Verbose uuid conflict logging, stream store utility fn migration.
